Resolve merge conflicts and issues from RB's router
