+++ packages/konsole-app/version.json
@@ -1,7 +1,7 @@
 {
   "build": {
-    "name":      "@dxos/konsole",
-    "buildDate": "2021-08-29T13:52:19.138Z",
+    "name":      "@dxos/konsole-app",
+    "buildDate": "2021-08-30T12:44:32.484Z",
     "version":   "1.2.9-alpha.18"
   }
 }
